test new cleanup time [swapfile]

+++ scripts/free-space.sh
@@ -7,6 +7,13 @@
 # dpkg-query -Wf '${Installed-Size}\t${Package}\n' | sort -n | tail -n 100
 # exit 0
 
+create_swapfile() {
+	sudo fallocate -l 4G /swapfile
+	sudo chmod 600 /swapfile
+	sudo mkswap /swapfile
+	sudo swapon /swapfile
+}
+
 if [ "${CI-false}" != "true" ]; then
 	echo "ERROR: not running on CI, not deleting system files to free space!"
 	exit 1
@@ -20,7 +27,6 @@ else
 	)
 
 	sudo apt purge -yq \
-		containerd.io \
 		snapd \
 		kubectl \
 		podman \
@@ -51,9 +57,15 @@ else
 	sudo rm -rf "/usr/local/share/boost"
 	sudo rm -rf "$AGENT_TOOLSDIRECTORY"
 
-	sudo docker image prune --all --force
-	sudo docker builder prune -a
+	# We shouldn't remove docker & it's images when running from `package_updates` workflow.
+	if [ "${CLEAN_DOCKER_IMAGES-true}" = "true" ]; then
+		sudo docker image prune --all --force
+		sudo docker builder prune -a
+		sudo apt purge -yq containerd.io
+	fi
 
 	sudo apt autoremove -yq
 	sudo apt clean
+
+	create_swapfile
 fi
